refactor(pages): use flat structure instead
