+++ src/pages/AssetDetailsPage.vue
