+++ src/pages/AssetsList.vue
