+++ src/pages/BlockDetailsPage.vue
