+++ src/pages/BlocksList.vue
